Add lock script validation in Tx's Output
A transaction's Output lock script will
now be syntactically validated.

Previously, only the lock+unlock pair was
validated. However this could lead to
loss of funds if the lock script is malformed.

This commit also fixes TxBuilder to copy
the same lock script as in the Tx that's
being attached to.

Previously TxBuilder wrongly assumed that
the previous Tx used a LockType.Key
and would interpret the first 32 bytes
as a public key.

The new validation routines catch this
as a syntactically invalid transaction.

Fixes #1380