Directly add our own enrollments to the pool when re-enrolling
Previously, we would not save our own enrollment to the pool when re-enrolling,
instead relying on the gossipping done by other nodes.

To make this more straightforward, we also split the 'EnrollmentPool.add' method
between add (with validation) and 'addValidated', in a manner similar to the Ledger.

Finally, the logic that prevents re-sending enrollments to other nodes have been eliminated.
Not re-sending the same data over and over again is a (useful) optimization,
however it was interfering with some tests (ValidatorRecurringEnrollment).
Moreover, we cannot trust that our Enrollment has made it to every validator
after we gossipped it, so it's better to flood it (for now).