Upgrade CI ubuntu runners to 20.04
Ubuntu 18.04 provides an older version of libsodium in it's
official repositories. We were previously using a PPA to get the
latest version but that seems to be either not present in the Github
runners or that PPA no longer provides libsodium.

Simplest way to get the most recent libsodium is to upgrade runners to
ubuntu 20.04