Encapsulate validator <-> index maps in their own structure
Those maps are slightly redundant (especially when it comes to the most recent height),
and problematic as they are currently stored in the EnrollmentManager.
Since they are validator-related, it would seem more appropriate to have them in
either the Ledger or the ValidatorSet. For such a transition to be easy,
it would be better if all methods related to it were properly encapsulated,
which this commit starts doing.
Further commits will move elements of the function inside the new structure.
This wasn't done in one go as this refactor also provides opportunity to
remove other usages, and to improve error handling (e.g. some parts depend
on the logger inside EnrollmentManager, which conflicts with our goal).