Remove redundant getCountOfValidators
EnrollmentManager.getCountOfValidators was only used in two places.
For both uses, it was following a pattern of getting the number of validators,
then looking up every entry to get the public key.
Since both usage only used the last known height, it can trivially be replaced
by using getActiveValidatorPublicKeys, which is a front to the ValidatorSet.
Once those two usages are simplified, the function could be removed.