Simplify preimage cache
`PreimageCycle` had a `populate` function that was only being used in
test code. This has been removed to simplify the code and testing of the
preimage cache code. Test code is added to check that the `seeds` and
`preimages` are correctly populated when the `PreimageCycle` is accessed
via the `opIndex` which takes a `Height` parameter. It also checks that
the nonce is correctly used beyond the first batch of preimages.