BanManager: Add the ability to weight reqs differently
Currently, we weight all requests the same.
However, some requests involve more than just one network requests.
An attacker could slow down a node by only failing the last request in such a scenario,
and delay other requests by as much as it's allowed.
To account for this scenario, we need to be able to pass this information to `onFailedRequest`.