Flash: Add a Negotiating stage to ChannelState
Previously the initial channel state was set to Rejected,
which is confusing because the channel counter-party
may not have been even contacted yet.

Instead, the initial phase has been changed to the new
'Negotiating' phase. This phase may transition to either
'SettingUp' or 'Rejected', based on the result of the
channel open exchange messages with the counter-party.