Ledger: Validate random seed inside acceptBlock
The 'check' in the validation routine was just doing a comparison with the provided value,
which the caller could do itself. Additionally the caller needs much more informations
to actually do the check (the current check is buggy), as it needs to derive the random
seed from the list of slashed validators.