Update validator well known keys
This is in preparation for issue 1248 which will enforce sorted inputs
and outputs within transactions. To keep the order of the freeze outputs
the same as the nodes used in unit tests the keys had to be updated to
have the same prefix so that the public key data was ascending.
The node keys now have bech32 addresses starting with `boa1xzval`.
e.g. boa1xzval2a3cdxv28n6slr62wlczslk3juvk7cu05qt3z55ty2rlfqfc6egsh2