Ledger.acceptBlock: Properly check for the random seed
The Ledger had two methods with similar name and use: getRandomSeed and getExternalizedRandomSeed.
The former was only used in Ledger.acceptBlock, while the later was used everywhere else the random
seed was needed. Upon closer inspection, it became obvious that 'getRandomSeed' was using
the presently-known Ledger state while getExternalizedRandomSeed had a much more predictable output.
In essence, 'getRandomSeed' could have been 'getCandidateRandomSeed'. This discrepancy could create
issues if a block which slashes a validator was externalized, but we received the pre-image in the
meantime, leading to a different random seed. With the removal of the former 'getRandomSeed' function,
we now have a predictable output (new state won't interfere).
In the process, the verbosely-name 'getExternalizedRandomSeed' was renamed to 'getRandomSeed'.