Nominator: Externalize block via the node's acceptBlock method
Before this change, the Nominator would directly externalize things through the Ledger.
The major downside of this approach is that any kind of special handling done by the
FullNode.acceptBlock (and Validator.acceptBlock) method is bypassed.
This became obvious then the stats were moved out of the Ledger and were not recorded anymore,
but it also means that the network whitelisting code of the FullNode was bypassed.