Correct random seed calculation
The random seed was incorrectly including `Hash.init` as the first item
hashed. It should only be the `preimage`s of the active validators.
The unit tests were also bypassing the testing of the random seed as the
tests did not have the `preimage`s available. Now there is a `preimage`
cache created using the `well known keypairs` which is used in the tests
to simulate `preimage`s being revealed.