Revamp Docker image build and push to support releases
With this change, the following will happen:
- The docker 'latest' tag, which is the default, will no longer be pushed to;
- When a PR is merged / something is pushed to a branch, 'devel' will be used instead;
- When a tag is pushed, the matching tag will be pushed to docker hub;
- When pushing a branch or a tag to one's fork, errors will no longer be generated;

The only downside here is to tie up the registry and agora build together,
so that if one fails, both have to be repeated. In practice, this almost never happened,
so it shouldn't be a great concern.

The goal is to latest trigger a push to 'latest' when a release is made.