unit test: simplify some node.Ledger tests
The TestLedger was using NODE2 which was making the tests more easily
broken when the order of the enrollments in the Test GenesisBlock was
updated. Now the TestLedger uses the first NODE in the list of sorted
enrollments.
Also the testing of missing pre-images was simplified to use helper
function `simulatePreimages` which can take a list of enrollment indexes
to skip (i.e. not reveal preimage).