Add signature and publickey in EncryptionKey
Votera app encrypts VoterCard and stores it in the app for ease of use.
When a user operates multiple nodes and receives an encryption key from Talos, it is not possible to determine which node the encryption key belongs to.

In addition, I think a safer voting system can be provided if a method is provided to verify whether the data received via QR code is accurately delivered.