Unit test: Use node of first enrollment in test
The test assumed that NODE2 was always the first node in the test
GenesisBlock. When there is a change which affects the utxo hashes of
the node's enrollments the order will change. In tests it is best to use
`agora.consensus.data.genesis.Test.genesis_validator_keys` to access the
keys in the order that they are in the GenesisBlock. When the order is
changed then a unit test in `agora.consensus.data.genesis.package` will
fail and print the correct order to update the array.