Network tests: Remove override of `prepareNominatingSet`
When testing we want to be as close to production behavior as possible.
This removes the txs_to_nominate config option and makes use of
mockClock only to set the time for nomination. This is much closer to
the real life case where nomination is triggered by time and not number
of transactions.