Flash: Fix spurious test failure
waitChannelState() used to be triggered by any node reaching the
specified state for a channel.

So, when we waitChannelState() for a chan_id/state and then
query a specific node, that node may not have reached that state yet.

By adding a PublicKey argument to waitChannelState() we can make sure
that the node in question actually reaches the state we want.