Network unit test: Prevent race condition failure
In `InvalidBlockSigByzantine` when we check the signature bitmask of the
last client we need to make sure that it has been given enough time to
have received all signatures from the other non-byzantine clients.
We can do this by waiting till the blocks are the same on our
non-byzantine clients before checking the bitmask.