Fix #2077: Refactor catch / discovery task to be DRYer
Those two functions could be refactored to have the timer start in the start method of the node.
We now only fire the discovery task immediately, as catchup relies on discovery
and is almost guaranteed to just return immediately.
Additionally this fixes a bug where Validators did not respect the `node.block_catchup_interval`
config field and always used 5 seconds.
Since we currently set this field to 2 seconds in the unittest,
this should significantly speed up tests that rely on catch-up.