Fold createRandomSeed functions together
Those functions were problematic for many reasons:
- The one in `EnrollmentManager` gets a list of `keys` but ignores missing ones;
- As a result, even though it ensures `keys.length > 0`, it can potentially call the one in `Block` with `count == 0`;
- The one in `Block` used `count`, which might consume the range, as it's an `InputRange`. `save` should have been used.
- It's a lot of duplication for a single `reduce` call;
- The need to have it in `Block` seems to come only from `makeTestBlock`, which needs to go,
  as the functionality needs too much context that only the Ledger can provide;

We are currently seeing the `count == 0` assert being spuriously triggered,
and this is the first step in fixing it.

Once that was done, EM.getRandomSeed was merged into Ledger.getRandomSeed,
as this ensures that there is only one place where we get the random seed from.
The confusing semantic of the method which allowed to provide
a list of nodes to ignore, but would ignore missing ones as well,
was changed to a regular exception, which we might spuriously see now.