`Tx` setters for `inputs` and `outputs`
In the mutable `Transaction` constructor we sort the inputs and outputs
and in the immutable `Transaction` constructor we assert if given
unsorted `inputs` and `outputs`. There was still an issue if the zero
arg constructor was used and then unsorted `inputs` or `outputs` were
set. This adds setters that assert if the lists are not valid when set.