Catch up with missing preimages regardless of nomination status
Previously, we were only catching up with pre-images we've seen as missing
through side effects of the `isInvalidPreimageRootReason`.
This `private` member of the `Ledger` was then cleared in the `NetworkManager`.

Instead of relying on this tight coupling and hard to follow logic,
and as we want to try to catch-up with all missing pre-images anyways,
we can just add a task in `Validator` which behaves similarly to the block catchup.