Make sure we give our PreImage to the ValidatorSet
Currently, we rely on some custom logic (such as hasRevealedPreimage
returning true for ourself) in a few places where we could just give
our pre-image to the ValidatorSet and remove special casing.
This is an attempt at doing that, making sure that the result of
'Ledger.getValidators' is usable (currently the pre-image is
not updated and thus the Ledger consider it a candidate for slashing).