`BitMask` `struct` for `validators`
The `BitMask` `struct` is specialized for use as the field in the
`BlockHeader` to record which validators have signed that block.
The number of active validators allowed to sign the block are set in the
constructor and can not be changed.
Trying to read or write to bits beyond the count is not allowed and will
cause an assertion.