Rename pre-image revelation function to improve tests
Modifying the behavior of the EnrollmentManager is problematic,
as other parts of the code may depend on it.
For example, the node currently special-case itself in `hasRevealedPreimage`
to work around this. If we want to avoid revealing pre-image,
just override the function that is supposed to reveal them,
not the function that provides them.