Remove a flakey test that relies on PreImage not being propagated
This test has an assert which checks that a node doesn't know its own pre-image,
and then that the API receive less than 100 (not 400, as documented) calls.
A node should always know its own pre-image, and as we are implementing better
gossiping for pre-image, the test will spuriously (or reliably) fail.
The very basis of this test (testing that an endpoint is called less than X times)
is dubious in itself, as it doesn't test an external behavior of the system.