SCP: Remove SecretKey type, files, and related string formatting functions
SCP doesn't need to know about the kind of `SecretKey`,
it can be handled solely by the D code.
However, along with `SecretKey.h` was the string-formatting ability for `NodeID`
(which in Stellar is a `PublicKey`), which we need to replace.
This is nowadays done via `SCPDriver.getStrKey`, however `QuorumIntersectionChecker`
does not have access to it, so it remains the last use of `StrKey` and its dependencies.