API: Introduce handshake, avoid relying on Exceptions
When 'getPublicKey' was first introduced, it was under the assumption that
a node would be either a FullNode or a Validator.
However, this had a major oversight: while a FullNode will never promote
to a Validator internally, from the outside, a node could switch between
a FullNode and a Validator (because an unenrolled Validator is no different
from a FullNode). This led to a few complications in the NetworkManager
and related code, including this hack with HTTPStatusException.
This is now partially fixed by moving the 'handshake' endpoint to FullNode.