Talos: Fold containers into the components
The architecture of Talos should make it easy, if not trivial, to add a page.
However doing so currently require adding tons of boilerplate in multiple package,
following an undocumented convention.
In order to make it easier to achieve this, we should at least make sure that
all modules that needs to be added for a new page can be in the same package
(or close to it). An obvious place to start folding things is containers,
which have almost a one-to-one mapping with components.