ManyBlocks test: update `conf.payment_period`
We now restrict the signature catchup to be only since the last payment
block. This means that if the blocks are externalized fast enough then
the `assertSameBlocks` can potentially fail as signatures will not be
the same on all nodes. To avoid this in this test we simply set the conf
value for `payment_period` to be more than the blocks we will generate.