Bug fix for setting `BitMask`
We were using the `|=` operator to do a bitwise `OR` of the original
`BitMask` and the new `BitMask` setting. This was causing problems
during signature catchup as we were updating the signature to match the
incoming combined signatures but then setting the signed bits from both
the local and remote `BitMasks`.
I tried to use `opAssign` but then had problems in code that does not
assign the `validators` `BitMask` during construction, for example
`data.Block.makeNewBlock` assigns the `BlockHeader` fields after default
`ctor` is called. To require only constructing `BlockHeader` with
validators BitMask included is an undesirable restriction.
Another possible solution would be to set the `length` and handle the
allocation of the `ubytes` during the `opAssign` when the `length` is
zero, but it doesn't feel right to be doing this outside of the `struct`
`ctors`.
So the chosen solution is to have a function named `copyFrom` which
updates an existing `BitMask` with the values from the provided
`BitMask`.