Fold 'agora.registry' into 'agora.node'
The registry always stood out in that it had a separate 'API' module
outside of the 'api' package. It also had 'main' and 'Server',
which mirrored the 'node' setup.

However, now that the registry is part of Agora itself,
a single package with 3 modules is of little use,
especially given the dependencies between those modules.

We can simply fold the 'Config' module into `agora.node.Config`,
which imports it anyway, and move 'API' to the `agora.api` package.
Only the `Server` is left, which fits in the `agora.node` package.

Note that in the process of moving the `Config` code,
the `validate` function was changed to use `ensure` instead of
`enforce` + `format`.