Log reason why block is not accepted
The reason was missing in some cases when a test failed during
`acceptBlock` as on the call to `validateBlock` the second time it will
not always fail again. To fix this the `acceptBlock` should return the
reason for failure and it can then be used in logging or assertion
message.