Fix Unit test `LocalTransaction`
There were a couple of issues with the second unit test in this file.
1) `PickyLedger` should be a `ValidatingLedger`.
2) `PickyValidator` node had a `validatingLedger` constructed in the
`Validator` class. Then the ledger of the `PickyValidator` was set to
a constructed `PickyLedger`. On some occasions the original ledger was
being used when the intention was to use the `PickyLedger`.
This was fixed by using dependency injection. This also makes the
`FullNode` ledger use DI to remove an undesirable check if it is a
`Validator` within it's construction.