Fix 2294: Always provide block height stats
Our stats are backed by an AA / map because it is useful to have some stats
not showing by default (e.g. when we log status code).
However, in the case of block stats, we always want to return them,
so we can simplify the code and call 'collect' with the struct directly.