getValidators: Return validators that aren't yet slashed
The intent of getValidators is to return validators that are active at a given height.
Currently, when a validator gets slashed, we set slashed_height to the height
at which it gets slashed, which means it could have signed that height had it
revealed a pre-image (and not be slashed).
As a result, we need to change a simple 'greated than' to 'greater or equal'
to ensure that we will get the right list of validators when looking up
already-externalized blocks.