Use fee rate per byte in TxBuilder
This enables the setting of a fee per byte rate for calculating the tx
fees in TxBuilder. If there is a very small refund then it is not added
as an output but implicitly included in the fees, where fees are the sum
of the input values minus the sum of the output values.