Introduce a module to throw an exception with a formatted message
It's common to need to format a string while throwing an exception.
But there are two downsides to this:
- The code is more verbose, as an import of `std.format` is needed,
  and the call becomes `throw new Exception(format("...", args))`;
- Creating the `Exception` itself allocates with the GC;
- The formatting allocates too;

For this reason, the `ensure` method provided by `agora.common.Ensure`
is statically allocated when the thread starts, and has a buffer for
allocation. This ensures that we will not allocate (even if the code
is not annotated as `@nogc`) and thus we can avoid triggerring GC
collections on throwing Exceptions.

The name was choosen to avoid conflict with `std.exception : enforce`
which provides a similar functionality without statically allocating.