Tests: Allow to set the default log level instead of just disabling logs
Instead of a boolean to enable / disable logging, allow to set the default log level.
This affects the main thread (which runs the test) as well as the individual nodes.
The benefit of this approach is to allow a more fine-grained control over logs
without having to rebuild the binary, as well as speeding up the tests when
running with logs disabled, as the nodes will neither format nor write them.