SCP: Introduce an intermediary function for slice -> std::string
Currently, the `std::string` returned by D to C++ is corrupted.
The reason why isn't clear - the bindings might be at fault,
or it *could* be an ABI bug, although the former is more likely.
Like any other issue in software engineering,
we can solve it by introducing another level of indirection.