Coinbase transactions only at each `payoutPeriod`
A `Coinbase` transaction is only created every payout block as defined
by `ConsensusParams.PayoutPeriod` number of blocks. Previously only
Validator payouts were using this interval, which meant that when adding
fees to transactions we would create Coinbase transactions every block
which paid to the Commons Budget. Now the fees for both validators and
the Commons Budget are accumulated until the payout block.