Update TxBuilder to support fees
When a `transaction` is created then the sum of the input values minus
the sum of the output values is the implied `fee` for the transaction.
The `fee` must be at least the transaction size in bytes multiplied by
the minimum fee rate as defined in the `Consensus` parameters. Note
that this commit enables the Faucet to use a deprecated version of
`sign` to prevent a breaking change.