Move `addPreimage` from EM to Ledger, remove `addPreimages`
As we are making the Ledger more usable, we should remove the dependency on `enrollment_manager`.
One of the thin wrapper the EM provide was `addPreimage`, which is now moved to the Ledger.
The `addPreimages` (plural) method had odd semantic (ignoring subsequent preimages
if one of the returned preimages was not accepted) and its mechanics were redundant
with what the `ValidatorSet` was doing for validation anyway, so it was simply removed.