Ledger: Merge slashing of UTXOSet / ValidatorSet together
This refactors the slashing to be performed in one single function,
instead of two functions called from the respective update methods.

The rationale for doing so is to prevent any slashed validator
to have an impact on the block, which means slashing needs to be done first.

Once that change was applied, it also was obvious that most of the functions
were redundant with one another, hence why they were merged.
This also allows to remove `Ledger.getMissingValidatorsUTXOs` which was here
only to deal with the duplication. Additionally, using `getValidators` and
a `filter` call makes this trivial. Finally, `getValidators used to be called
needlessly (when there was no slashing informations), and this was fixed as well.