Change type of NodeID to ulong
SCP statements previously contain a `node_id` which is 64-byte key.
However because we are using staking we had the opportunity to
optimize that and reduce the network load. So we replace `node_id`
with a 8-byte `ulong` index. This index could be used to look up
the frozen UTXO of a validator for a specified height.