Make the name registry a proper DNS server
The function of the name registry was very similar to that of a DNS server.
However, it was working at a higher level, which means the clients depended
on both the DNS server and the registry working.
Combining both reduces the chances of error.

Additionally, this will allow a few key improvements:
- Nodes will be able to get an automatic domain name, e.g. `IDENTIFIER.net.bosagora.io`,
  where `IDENTIFIER` is an encoded version of their UTXO hash;
- This will in turn allow to automatically enable HTTPS for the admin interface,
  allowing a simple way for node operators to connect to their nodes from anywhere;
- It will integrate with the existing DNS seed, for which we have a client but no server;
- We get replication for free, as we can setup a slave DNS server which has no knowledge of Agora;