Reset `enrolls_keys_for_unknown_preimages` in Ledger
This set was only ever populated and never cleared so it would end up
checking all nodes for preimages even when it had the preimages already.
Also try to get all missing preimages from other nodes not just the ones
not missing in a consensus data from another node which we are missing.
The test for preimage sharing was also updated to only block preimage
sending from the nodes and allow retrieval of any preimages the node
has.