Implement an automatic config filler instead of doing it manually
Our configuration started pretty simple, and organically grew over time.
Originally, we wrote code to read the YAML `Node` and fill the `struct` / arrays / value,
as it was the most straightforward and efficient approach.
However, our config file has now reached a size where there is a lot of boilerplate
and repetition needed to achieve this.

Thus, this changes the implementation of the configuration to generate code from the type definition.
`parseConfig` will now walk the provided `struct` recursively, attempt to read each field,
use the default value if a sensible one is provided, or throw otherwise.

As a result, a few attributes needed to be added to the `Config` et al. type definitions
to inform the config filler of which fields were optional despite not having a different default,
as well as how to convert some fields (most notably `Duration`).

To support this change, `agora.common.Config` now only contains the config filling code,
with the type definitions having been moved to `agora.node.Config`.
To reduce dependencies, a new module, `agora.common.ConfigAttributes` has also been added.