Ledger: Reject TXs with lower fees compared to the pool
This will allow to protect against TX flood attacks.

With double_spent_threshold_pct and max_fee_undercut_pct, for an
attacker to flood more and more transactions to the network they would
have to pay increasingly more fees.

With double_spent_threshold_pct of %20, an attacker would have to pay almost 80M times
more than the average TX in the pool to be able to create 100 double-spent TXs.

With this, we can keep our pool size unlimited, but make it unpractical to use it as a
DoS vector.

Fixes #1959