Config: Implement built-in `core.time.Duration`
Duration is such an essential part of the config that having to add a converter
every time becomes a chore. Also, we end up relying on odd conventions,
such as some fields being suffixed with a `_secs` while other assume seconds.
Finally, specifying a duration in a fixed unit isn't very convenient for the user.
With the new approach, when a duration in a fixed unit is wanted,
a suffix can be used, while the default behavior (accepting units in mapping)
should be prefered when the duration range is much greater.