API: Rename POST endpoints to be consistent
We had a mix of PUT, POST, and 'receive_' endpoints, despite them having a similar functionality.
The 'receive' name was taken from SCP's 'receiveEnvelope' and is still used in the Nominator,
but the REST API endpoints are now all named using the same patternL `POST /type`.

`POST` is more adequate than `PUT` here, as `PUT` should refer to a specific resource
e.g. an endpoint that puts a signature by a specific pubkey would be adequately named
'PUT /public_key_signature/PUBLICKEY' if 'PUBLICKEY' was part of the URL,
but if the public key and signatures were both in the payload, the proper name
would be 'POST /public_key_signature' as it's a generic way to process a payload.

In our scheme, we could have 'PUT /transaction/TXHASH' but we'd need to validate
the hash is correct anyways, so it would only make usage of the endpoint harder.

For backward compatibility with Faucet, a `PUT /transaction` endpoint is still available,
but will be removed as soon as Faucet has been updated.