Nominator: Replace global ManagedDatabase lock with local one
The usage of ManagedDatabase predates the unification into stateDB / cacheDB.
Nowadays, we do not need such a global lock, as there are only two databases,
and their relation is well defined.
The envelope store stands out as a user of the cache DB and is a simple
place to start as the other users (e.g. Ledger) are more entangled.