SCPEnvelopeStore: Move to protocol and simplify variable name
This class is only used within the validator, so it could even be made package.
But at the very least, it should leave next to the Nominator,
as it feels out of place in the root consensus package,
and also makes it hard to exclude.

In the process of doing that move, the name of the variable in Nominator
was also shortened, as repeating the class name makes things quite verbose.