Add a class for block rewards calculations
Class `Reward` is able to calculate the rewards to the `Commons Budget`
and the `Validators` as defined in the `Bosagora Whitepaper`.
It also includes adjusting the total paid to the `Validators` when not
all blocks are signed by all `Validators`. This is to encourage all to
sign and gossip the block signatures of others.