Registry: Add a configuration file
As the role of the registry is expanding, it became needed to add a config file.

This will allow us to configure things that would be hard, verbose and awkward to
do via CLI arguments, such as the list of zones for which the server is authoritative,
or the logging configuration.