Registry: Introduce guessAddressType
This function can be used to 'guess' an address type from a string.
Ideally we should move to have typed input (e.g. PUT /addresses),
but as string is the common denominator, and we can fairly accurately
derive the type of of address from the string (the formats are mutually exclusive),
this is the path of least resistance as of now.