Flash: Drop ThinFlashNode & AgoraFlashNode
We now only have a single Flash node class called FlashNode.
Anything related to a agora node is handled by delagates, so we still
can have a ThinFlashNode like operation.

This also makes use of the getFlashNode & putFlashNode endpoints of the
NameRegistry. Nodes now need a open channel to be able to register themselves.
In case a node has no open channels, they pass their address to their peer as
they are openning their first channel to bootstrap the communication.