Store the fees and data_fees for each height
Fees are now only paid to Validators for blocks they sign. The `Coinbase`
transaction is now included in the block at the end of the next payout period.
The fees are the `sum of the inputs` minus the `sum of the outputs`.
If we do not store the fees when we externalise a block then some inputs may be
spent before the payout is made. The data fees are also stored as they are only
paid to the 'Commons Budget' so are subtracted from the fees when paying the
validators their percentage of fees.