Registry: Document and flatten configuration
The registry configuration was a bit lacking, and this has been fixed.
Additionally, since the whole point of the 'registry' is now to have an 'enabled' boolean,
and then contain the DNS configuration, the two configurations were flattened.
One reason to do so is that we want to move towards the registry being mostly DNS-based
(as opposed to mostly HTTP-based, as currently), if not only DNS-based.
Hence, the registry and the DNS server are now tied together.
The original reason for allowing disabling the DNS server was to not impair
production if a bug was found in the DNS server, but this concern is no more.