Registry: Rework configuration to allow non-authoritative servers
This rework two main aspects of the configuration:
- It allows non-authoritative server, by making 'authoritative' a boolean of the zone;
- It restricts zone configuration to the currently envisioned zone split;

The second point warrant more explanation: previously, the DNS server was developed to be generic,
and thus allowed to declare arbitrary zone names in the configuration.
However, that wasn't yet used by the node. As we are preparing to use it, one fact came up:
the node will need to have a matching configuration, that is, know what domain to look the validators
and the flash nodes in. Worse, this flexibility was actually not usable,
as only enrolled nodes (or nodes with a channel) could actually register.

Hence, this change things to make client configuration much easier, with no expense for the DNS server:
validators now live under 'validators.realm', while flash nodes are found under 'flash.realm'.
Realm names are expected to be 'testnet.bosagora.io', 'coinnet.bosagora.io', and 'localhost' (for local dev).