Registry: Refactor zone / domain parsing
The original idea was that the same name would be accessible under different zones.
This proved to be unnecessary, and so having the iteration is of little use.
Additionally, as we are implementing zone transfer, there is a need for the code
in 'matches' to be moved to its own function.