Tests: Reduce amount of imports boilerplate
Writing a tests historically required a lot of boilerplate,
but this was gradually trimmed down as our test framework evolved.
This is another attempt at trimming it down by removing a lot of
no-longer-needed imports, as well as adding common ones to Base
as public imports (e.g. almost every tests use 'Transaction').