Registry: Always use qname with CNAME record
The original code came from a misinterpretation of the spec:
it says 'copy [...], change QNAME [...], and go back [...]'.

But those things should not happen simultaneously, so the copy should happen
with the proper QNAME, then the algorithm should be restarted with a new QNAME.
However, this is done to support A records in the same domain, e.g. when
'a.example.com' is a CNAME for 'b.example.com' and the latter has an 'A' record.

Currently, 'PUBKEY.validators.REALM' cannot point to 'XXXX.validators.REALM',
as it could only point to another pubkey. It could in the future, relying on
domain matching to select the right key, but this first require us to fix
the validation of record added to the registry.