Simplify block creation by adding a Ledger function
The Ledger has all the context required to create a Block.
Instead of calling the free function, we can offer the use the ability
to create a block via this function, which simplifies the Nominator
and will simplify tests. It also make using the Ledger in library code
much easier.
The free function stay useful for users that do not instantiate a Ledger.