BlockHeader: Replace 'missing_validators' and 'random_seed' with 'preimages'
The original idea was that we wouldn't need to store the individual pre-images in the block,
and missing_validators would seldom be populated.
However, we ended up needing the individual pre-images, and we currently have multiple
issues related to catchup with unknown pre-images.
This definitely fixes the problem by storing the pre-images in the block.
The downside of this approach is that the BlockHeader's size is now linearly
proportional to the number of validators.
However, as validators are expected to be long-lived, a future optimization would
be to remove earlier pre-images from storages when a node publishes new pre-images,
essentially reducing the storage to one `Hash` per unique enrollment.