Replace % with percentage in log messages
When % is used in a log message then it would need to be escaped to
prevent crash at runtime. To be on the safe side I have opted to replace
`%` with the text `percentage` to avoid any issues as it is not caught
by unit tests as they do not write the message to file.